refactor CustomizeHunctionSetting.jsx
